fish: Print prompt stuff before prompt
Previously I was printing a bunch of stuff (e.g. current directory, git
branch, exit code) in the `fish_prompt` function itself. This worked,
but when a terminal got too narrow, fish would replace the whole prompt
with the short default `>`:
https://github.com/fish-shell/fish-shell/issues/904

As suggested somewhere in that issue, I'm now printing the first line of
the prompt (with all the extra info) in a separate function that runs on
the `fish_prompt` event. The actual prompt is just the second line, a
single `$ `.