Go back to original prompt, but shorten it
In the last commit, I changed how the fish prompt was generated. I ended
up not liking this, since it was noticeably slower to render. Now, just
shortening the prompt to remove the word "exited" when the last command
failed. It'll just show a number in red. This will usually be fine; the
prompt won't end up long enough to be truncated.